Prepare for mobile use
